requirements.txt
pytest
selenium
requests
pytest-xdist
pyyaml